+++ src/db/migrations/v1.4.0-add-council.js
@@ -1,0 +1,13 @@
+"use strict";
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    return queryInterface.addColumn(
+      "registrations",
+      "council",
+      Sequelize.STRING
+    );
+  },
+  down: queryInterface => {
+    return queryInterface.removeColumn("registrations", "council");
+  }
+};